Replace budget overview toast with a real modal.
Period expense total now opens traveller budget breakdown in a dismissible modal so overview matches feedback-style UX and drops the budgetOverview toast path.
