+++ TravelCostApp/components/ExpensesOutput/ExpensesSummary.tsx
@@ -1,5 +1,5 @@
 import { Platform, StyleSheet, View, Text } from "react-native";
-import React, { useCallback, useContext, useEffect, useState } from "react";
+import React, { useContext, useState } from "react";
 import { GlobalStyles } from "../../constants/styles";
 import { shadowRegressionStyles, periodHeaderLabelFontSize } from "../../styles/shadow-regression-styles";
 import * as Progress from "react-native-progress";
@@ -7,14 +7,12 @@ import { TripContext } from "../../store/trip-context";
 
 import { i18n } from "../../i18n/i18n";
 
-import { formatExpenseWithCurrency, truncateNumber } from "../../util/string";
 import PropTypes from "prop-types";
 import { Pressable } from "react-native";
 import Toast from "react-native-toast-message";
 import { MAX_JS_NUMBER } from "../../confAppConstants";
 import * as Haptics from "expo-haptics";
 import { UserContext } from "../../store/user-context";
-import { getRate } from "../../util/currencyExchange";
 import { SettingsContext } from "../../store/settings-context";
 import {
   ExpenseData,
@@ -28,6 +26,7 @@ import {
   calculateDailyAverage,
   getBudgetColor,
 } from "../../util/budget";
+import BudgetOverviewModal from "./BudgetOverviewModal";
 
 const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
   const tripCtx = useContext(TripContext);
@@ -35,23 +34,7 @@ const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
   const expCtx = useContext(ExpensesContext);
   const { settings } = useContext(SettingsContext);
   const hideSpecial = settings.hideSpecialExpenses;
-  const [isToastShowing, setIsToastShowing] = useState(false);
-
-  const [lastRate, setLastRate] = useState(1);
-  const lastRateUnequal1 = lastRate !== 1;
-  const getRateCallback = useCallback(async () => {
-    if (!userCtx.lastCurrency || !tripCtx.tripCurrency) {
-      setLastRate(1);
-      return;
-    }
-    setLastRate(await getRate(tripCtx.tripCurrency, userCtx.lastCurrency));
-  }, [tripCtx.tripCurrency, userCtx.lastCurrency]);
-  useEffect(() => {
-    async function call() {
-      await getRateCallback();
-    }
-    call();
-  }, [userCtx.lastCurrency, tripCtx.tripCurrency, getRateCallback]);
+  const [isOverviewVisible, setIsOverviewVisible] = useState(false);
 
   const safeExpenses = Array.isArray(expenses) ? expenses : [];
   const travellers = Array.isArray(tripCtx.travellers)
@@ -82,23 +65,9 @@ const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
     return <></>;
   }
 
-  const expensesSumString = formatExpenseWithCurrency(
-    truncateNumber(expensesSum, 1000, true),
-    tripCurrency,
-  );
-
-  const calcExpensesSumString =
-    lastRate == 1
-      ? ""
-      : formatExpenseWithCurrency(
-          truncateNumber(expensesSum * lastRate, 1000, true),
-          userCtx.lastCurrency,
-        );
-
   let budgetNumber = Number(tripCtx.dailyBudget ?? 0);
   let infinityString = "";
   let periodExpenses: ExpenseData[] = [];
-  let periodLabel = "";
   const expenseSumNum = Number(expensesSum);
   let totalBudget = Number(tripCtx.totalBudget ?? 0);
   if (Number.isNaN(totalBudget)) totalBudget = 0;
@@ -108,30 +77,25 @@ const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
   switch (periodName) {
     case "day":
       periodExpenses = expCtx.getRecentExpenses(RangeString.day) || [];
-      periodLabel = i18n.t("todayLabel");
       break;
     case "week":
       budgetMult = 7;
       budgetNumber = budgetNumber * budgetMult;
       periodExpenses = expCtx.getRecentExpenses(RangeString.week) || [];
-      periodLabel = i18n.t("weekLabel");
       break;
     case "month":
       budgetMult = 30;
       budgetNumber = budgetNumber * budgetMult;
       periodExpenses = expCtx.getRecentExpenses(RangeString.month) || [];
-      periodLabel = i18n.t("monthLabel");
       break;
     case "year":
       budgetMult = 365;
       budgetNumber = budgetNumber * budgetMult;
       periodExpenses = expCtx.getRecentExpenses(RangeString.year) || [];
-      periodLabel = i18n.t("yearLabel");
       break;
     case "total":
       budgetNumber = totalBudget ?? MAX_JS_NUMBER;
       periodExpenses = expCtx.expenses || [];
-      periodLabel = i18n.t("totalLabel");
       break;
     default:
       break;
@@ -199,113 +163,49 @@ const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
   if (Number.isNaN(budgetProgress)) {
     return <></>;
   }
-  const calcLeftToSpend = lastRateUnequal1
-    ? formatExpenseWithCurrency(
-        truncateNumber((budgetNumber - expenseSumNum) * lastRate, 1000, true),
-        userCtx.lastCurrency,
-      )
-    : "";
-  const leftToSpendString = `${i18n.t(
-    "youHaveXLeftToSpend1",
-  )}:\n${formatExpenseWithCurrency(
-    truncateNumber(budgetNumber - expenseSumNum, 1000, true),
-    tripCurrency,
-  )}${lastRateUnequal1 ? " = " : ""}${calcLeftToSpend}${i18n.t(
-    "youHaveXLeftToSpend2",
-  )}`;
-
-  const calcOverBudget = lastRateUnequal1
-    ? formatExpenseWithCurrency(
-        truncateNumber((expenseSumNum - budgetNumber) * lastRate, 1000, true),
-        userCtx.lastCurrency,
-      )
-    : "";
-
-  const overBudgetString = `${i18n.t(
-    "exceededBudgetByX1",
-  )}:\n${formatExpenseWithCurrency(
-    truncateNumber(expenseSumNum - budgetNumber, 1000, true),
-    tripCurrency,
-  )}${lastRateUnequal1 ? " = " : ""}${calcOverBudget} !`;
-
-  const periodBudgetString = `${periodLabel} ${i18n.t(
-    "budget",
-  )} :\n${formatExpenseWithCurrency(
-    budgetNumber,
-    tripCtx.tripCurrency,
-  )} = ${formatExpenseWithCurrency(
-    budgetNumber * lastRate,
-    userCtx.lastCurrency,
-  )}`;
-
   const valid = tripCtx.tripid && travellerNames.length > 0;
+  const closeOverview = () => setIsOverviewVisible(false);
+
   const pressBudgetHandler = () => {
-    if (isToastShowing) {
-      setIsToastShowing(false);
-      Toast.hide();
+    if (isOverviewVisible) {
+      closeOverview();
       return;
     }
-    setIsToastShowing(true);
     Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
     if (infinityString) {
       Toast.show({
         type: "error",
         text1: i18n.t("noTotalBudget"),
         text2: i18n.t("infinityLeftToSpend"),
-        onHide() {
-          setIsToastShowing(false);
-        },
       });
       return;
     }
-    const tripCurrency = tripCtx.tripCurrency;
-    const lastCurrency = userCtx.lastCurrency;
     if (!valid) {
-      Toast.hide();
-      setIsToastShowing(false);
       return;
     }
-    Toast.show({
-      type: "budgetOverview",
-      position: "bottom",
-      text1: `${periodLabel} ${i18n.t("expenses")} :\n${expensesSumString} ${
-        lastRateUnequal1 ? "=" : ""
-      } ${calcExpensesSumString}`,
-      text2:
-        budgetNumber > expenseSumNum
-          ? `${leftToSpendString}`
-          : `${overBudgetString}`,
-      bottomOffset: 40,
-      visibilityTime: 20000,
-      onHide: () => {
-        setIsToastShowing(false);
-      },
-      props: {
-        text3: `${formatExpenseWithCurrency(
-          1,
-          tripCurrency,
-        )} = ${formatExpenseWithCurrency(lastRate, lastCurrency)}`,
-        travellerList: travellerNames,
-        travellerBudgets:
-          travellerNames.length > 0 ? budgetNumber / travellerNames.length : 0,
-        budgetNumber: budgetNumber,
-        travellerSplitExpenseSums: travellerSplitExpenseSums,
-        currency: tripCurrency,
-        noTotalBudget: noTotalBudget,
-        periodName: periodName,
-        periodLabel: periodLabel,
-        periodBudgetString: periodBudgetString,
-        lastRateUnequal1: lastRateUnequal1,
-        trafficLightActive: settings.trafficLightBudgetColors,
-        currentBudgetColor: budgetColor,
-        averageDailySpending: averageDailySpending,
-        dailyBudget: dailyBudget,
-        expenseSumNum: expenseSumNum,
-      },
-    });
+    setIsOverviewVisible(true);
   };
 
   return (
+    <>
+    <BudgetOverviewModal
+      isVisible={isOverviewVisible}
+      onClose={closeOverview}
+      travellerList={travellerNames}
+      travellerBudgets={
+        travellerNames.length > 0 ? budgetNumber / travellerNames.length : 0
+      }
+      travellerSplitExpenseSums={travellerSplitExpenseSums}
+      currency={tripCurrency}
+      noTotalBudget={noTotalBudget}
+      periodName={periodName}
+      trafficLightActive={settings.trafficLightBudgetColors}
+      currentBudgetColor={budgetColor}
+      averageDailySpending={averageDailySpending}
+      dailyBudget={dailyBudget}
+      expenseSumNum={expenseSumNum}
+      budgetNumber={budgetNumber}
+    />
     <Pressable
       testID="expenses-summary-pressable"
       onPress={() => pressBudgetHandler()}
@@ -335,6 +235,7 @@ const ExpensesSummary = ({ expenses, periodName, style = {} }) => {
         height={constantScale(6, 0.5)}
       />
     </Pressable>
+    </>
   );
 };
